requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
bcrypt==3.2.2
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.2
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fast_captcha==0.3.2
fastapi==0.116.1
fastapi-limiter==0.1.6
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.4
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.25.0
google-cloud-texttospeech==2.16.2
google-generativeai==0.5.4
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.62.3
gTTS==2.5.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
ipython==8.26.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
limits==5.4.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.14.2
openai-whisper==20250625
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt-toolkit==3.0.47
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psycopg[binary]==3.2.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-multibase==1.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.2.3
pytest==8.4.1
pytest-env==1.1.3
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2025.7.34
reportlab==4.2.0
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.42
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.2
stripe==9.6.0
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.17.2
yarl==1.20.1
